+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/Engine.kt
@@ -5,6 +5,7 @@
 package mozilla.components.concept.engine
 
 import android.content.Context
+import android.os.Parcelable
 import android.util.AttributeSet
 import android.util.JsonReader
 import androidx.annotation.MainThread
@@ -220,6 +221,17 @@ interface Engine : WebExtensionRuntime, DataCleanable {
     fun registerServiceWorkerDelegate(
         serviceWorkerDelegate: ServiceWorkerDelegate
     ): Unit = throw UnsupportedOperationException("Service workers support not available in this engine")
+
+    /**
+     * Handles user interacting with a web notification.
+     *
+     * @param webNotification [Parcelable] representing a web notification.
+     * If the `Parcelable` is not a web notification this method will be no-op.
+     *
+     * @see <a href="https://developer.mozilla.org/en-US/docs/Web/API/Notification">MDN Notification docs</a>
+     */
+    fun handleWebNotificationClick(webNotification: Parcelable): Unit =
+        throw UnsupportedOperationException("Web notification clicks not yet supported in this engine")
 
     /**
      * Fetch a list of trackers logged for a given [session] .